#把所有的decimal.ToInt16()改成ToInt32();这样为int而非short；+新增右键追加步和删除步；#修改保存工程的封装Value数据的方法体，去掉用IndexOf取StepIndex，改为for循环；

+++ LightController/Ast/Entity/KeyEntity.cs
@@ -26,8 +26,8 @@ namespace LightController.Entity
             this.Key1Array = new string[24];
             for (int i = 0; i < 24; i++)
             {
-                this.Key0Array[i] = StringHelper.DecimalStringToBitHex(Convert.ToInt16(data[i]).ToString(), 2);
-                this.Key1Array[i] = StringHelper.DecimalStringToBitHex(Convert.ToInt16(data[i + 24]).ToString(), 2);
+                this.Key0Array[i] = StringHelper.DecimalStringToBitHex(Convert.ToInt32(data[i]).ToString(), 2);
+                this.Key1Array[i] = StringHelper.DecimalStringToBitHex(Convert.ToInt32(data[i + 24]).ToString(), 2);
             }
         }
