+++ LightController/MyForm/Multiplex/MultiStepCopyForm.cs
@@ -112,8 +112,8 @@ namespace LightController.MyForm
 		private void enterButton_Click(object sender, EventArgs e)
 		{
 			// 1. 起始步、结束步的验证
-			int startNum = Decimal.ToInt16(startNumericUpDown.Value);
-			int endNum = Decimal.ToInt16(endNumericUpDown.Value);
+			int startNum = Decimal.ToInt32(startNumericUpDown.Value);
+			int endNum = Decimal.ToInt32(endNumericUpDown.Value);
 			if (startNum > endNum)
 			{
 				MessageBox.Show("起始步不可大于结束步；请重新选择。");
